requirements.txt
# Discord Bot
discord.py==2.3.2
python-dotenv==1.0.0

# Data Collection
feedparser==6.0.10
sec-edgar-downloader==5.0.3
yfinance==0.2.32
fredapi==0.5.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# NLP & ML
transformers==4.35.2
torch==2.1.1
sentencepiece==0.1.99
spacy==3.7.2

# Database
aiosqlite==0.19.0

# Utils
pandas==2.1.3
numpy==1.26.2
python-dateutil==2.8.2
pytz==2023.3
APScheduler==3.10.4